rename: taskdetail 폴더 생성 후 컴포넌트 이동
